Refactored Database project
