+++ README.md
@@ -59,25 +59,13 @@ $ npm install -g gulp
 
 ### Add database
 
-> Before adding the database, make sure you have setup your configuration correctly. 
-> If you have not added a command line argument for the database connection string then it will be resolved from the configuration json instead.
-
 Run in the `src/SFA.DAS.EmployerFinance.Database` directory:
 
 ```powershell
 > dotnet run
 ```
-Or to run with a custom database connection string:
 
-```powershell
-> dotnet run <database connection string>
-```
-
-* Wait until you see the following in the shell window and then press Ctrl + C:
-  ``` 
-  [SFA.DAS.EmployerFinance.Database.EmployerFinanceDatabaseHelper] - Finished deploying database 
-  ```
-  
+* Wait until you see `[SFA.DAS.EmployerFinance.Database.Jobs.DeployDatabaseJob] - Finished deploying database ` in the shell window and then press Ctrl + C.
 
 ### Add packages
