+++ src/SFA.DAS.EmployerFinance.Database/Program.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using Microsoft.Extensions.Logging;
 using SFA.DAS.EmployerFinance.Database.DependencyResolution;
+using SFA.DAS.EmployerFinance.Database.Jobs;
 using SFA.DAS.EmployerFinance.Extensions;
 
 namespace SFA.DAS.EmployerFinance.Database
